fix: 잘못된 폴더로 들어가 있어 수정
