--- src/main/java/zoo/insightnote/domain/comment/dto/request/CommentDefaultResponse.java
@@ -1,15 +1,0 @@
-package zoo.insightnote.domain.comment.dto.request;
-import java.time.LocalDateTime;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.AccessLevel;
-
-public record CommentDefaultResponse (
-     Long commentId,
-     String content,
-     LocalDateTime createAt,
-     String author
-){
-}
