--- src/main/java/zoo/insightnote/domain/comment/dto/request/CommentIdResponse.java
@@ -1,5 +1,0 @@
-package zoo.insightnote.domain.comment.dto.request;
-
-
-public record CommentIdResponse(Long id) {
-}
