+++ src/main/java/zoo/insightnote/domain/comment/dto/response/CommentDefaultResponse.java
@@ -1,0 +1,10 @@
+package zoo.insightnote.domain.comment.dto.response;
+import java.time.LocalDateTime;
+
+public record CommentDefaultResponse (
+     Long commentId,
+     String content,
+     LocalDateTime createAt,
+     String author
+){
+}
