+++ src/main/java/zoo/insightnote/domain/comment/dto/response/CommentIdResponse.java
@@ -1,0 +1,5 @@
+package zoo.insightnote.domain.comment.dto.response;
+
+
+public record CommentIdResponse(Long id) {
+}
