+++ src/main/java/zoo/insightnote/domain/comment/service/CommentService.java
@@ -6,11 +6,11 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import zoo.insightnote.domain.comment.dto.request.CommentIdResponse;
-import zoo.insightnote.domain.comment.dto.request.CommentListResponse;
+import zoo.insightnote.domain.comment.dto.response.CommentIdResponse;
+import zoo.insightnote.domain.comment.dto.response.CommentListResponse;
 
-import zoo.insightnote.domain.comment.dto.response.CommentCreateRequest;
-import zoo.insightnote.domain.comment.dto.response.CommentUpdateRequest;
+import zoo.insightnote.domain.comment.dto.request.CommentCreateRequest;
+import zoo.insightnote.domain.comment.dto.request.CommentUpdateRequest;
 
 import zoo.insightnote.domain.comment.entity.Comment;
 import zoo.insightnote.domain.comment.mapper.CommentMapper;
